reorganise folders and functions
